\name{ppmvb-package}
\alias{ppmvb-package}
\docType{package}
\title{
Variational Bayes version of the \code{\link{ppm}} function in \link{spatstat}
}
\description{
Use variational Bayes approximation of logistic regression to fit exponential family Gibbs models.
}
\details{
The main function is \code{\link{ppmvb}} which should work just like \code{\link{ppm}}.
}
\author{
Tuomas Rajala <tuomas.rajala@iki.fi>
}
\keyword{ package }
\seealso{
\code{\link{spatstat}}
}
